Avanzamento

+++ jar/src/main/java/lispa/schedulers/dao/sgr/sire/history/SireHistoryWorkitemDAO.java
@@ -613,7 +613,7 @@ public class SireHistoryWorkitemDAO {
 			throws Exception {
 		ConnectionManager cm = null;
 		Connection connection = null;
-		List<Integer> listPk = new ArrayList<>();
+		List<Long> listPk = new ArrayList<>();
 		try {
 			cm = ConnectionManager.getInstance();
 			connection = cm.getConnectionOracle();
@@ -633,7 +633,8 @@ public class SireHistoryWorkitemDAO {
 
 				try (ResultSet rs = ps2.executeQuery();) {
 					while (rs.next()) {
-						listPk.add(rs.getInt(1));
+						
+						listPk.add(rs.getLong(1));
 					}
 					int numberInParameters = listPk.size();
 					String parameters = "";
@@ -652,7 +653,7 @@ public class SireHistoryWorkitemDAO {
 						// Array array=connection.createArrayOf("NUMBER",
 						// listPk.toArray());
 						for (int i = 0; i < listPk.size(); i++) {
-							ps.setInt(i + 2, listPk.get(i));
+							ps.setLong(i + 2, listPk.get(i));
 						}
 						ps.setTimestamp(listPk.size() + 2, DataEsecuzione
 								.getInstance().getDataEsecuzione());
